refactor folders styles
