+++ client/src/components/HeroCarousel.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Carousel } from 'react-responsive-carousel';
 import "react-responsive-carousel/lib/styles/carousel.min.css"; // Import the carousel styles
-import '../pages/LandingPage.css'; // Your existing styles
+import '../pages/styles/LandingPage.css'; // Your existing styles
 
 const HeroCarousel = () => (
   <section className="carousel">
