+++ client/src/pages/AboutPage.jsx
@@ -1,6 +1,6 @@
 // AboutPage.jsx
 import React from "react";
-import "./AboutPage.css";
+import "./styles/AboutPage.css";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
